chore: add verbatimModuleSyntax test case
