+++ crates/pack-tests/tests/snapshot/basic/foreign_jsx_transform/input/comp.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Comp = () => <div>Comp</div>
+
+export default Comp;
