+++ crates/pack-tests/tests/snapshot/basic/foreign_jsx_transform/input/index.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import 'foreign-jsx-transform';
+import Comp from './comp';
 
 const App = () => {
-    return <div></div>
+    return (
+        <div>
+            <Comp></Comp>
+        </div>
+    )
 }
 
 export default App;
