--- crates/pack-tests/tests/snapshot/basic/foreign_jsx_transform/output/_project__basic_foreign_jsx_transform_32e9a332.js
@@ -1,46 +1,0 @@
-(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([typeof document === "object" ? document.currentScript : undefined,
-"[project]/basic/foreign_jsx_transform/node_modules/react/index.js [client] (ecmascript)", ((__turbopack_context__) => {
-"use strict";
-
-const React = {
-    createElement: function createElement() {},
-    Fragment: 'Fragment',
-    Component: function Component() {}
-};
-var __TURBOPACK__default__export__ = React;
-const createElement = React.createElement;
-const Fragment = React.Fragment;
-const Component = React.Component;
-__turbopack_context__.s([
-    "default",
-    0,
-    __TURBOPACK__default__export__
-]);
-}),
-"[project]/basic/foreign_jsx_transform/node_modules/foreign-jsx-transform/index.js [client] (ecmascript)", ((__turbopack_context__) => {
-"use strict";
-
-/** @jsx jsx */ var a = 1;
-var __TURBOPACK__default__export__ = a;
-__turbopack_context__.s([]);
-}),
-"[project]/basic/foreign_jsx_transform/input/index.jsx [client] (ecmascript)", ((__turbopack_context__) => {
-"use strict";
-
-var __TURBOPACK__imported__module__$5b$project$5d2f$basic$2f$foreign_jsx_transform$2f$node_modules$2f$react$2f$index$2e$js__$5b$client$5d$__$28$ecmascript$29$__ = __turbopack_context__.i("[project]/basic/foreign_jsx_transform/node_modules/react/index.js [client] (ecmascript)");
-var __TURBOPACK__imported__module__$5b$project$5d2f$basic$2f$foreign_jsx_transform$2f$node_modules$2f$foreign$2d$jsx$2d$transform$2f$index$2e$js__$5b$client$5d$__$28$ecmascript$29$__ = __turbopack_context__.i("[project]/basic/foreign_jsx_transform/node_modules/foreign-jsx-transform/index.js [client] (ecmascript)");
-;
-;
-const App = ()=>{
-    return /*#__PURE__*/ __TURBOPACK__imported__module__$5b$project$5d2f$basic$2f$foreign_jsx_transform$2f$node_modules$2f$react$2f$index$2e$js__$5b$client$5d$__$28$ecmascript$29$__["default"].createElement("div", null);
-};
-var __TURBOPACK__default__export__ = App;
-__turbopack_context__.s([
-    "default",
-    0,
-    __TURBOPACK__default__export__
-]);
-}),
-]);
-
-//# sourceMappingURL=_project__basic_foreign_jsx_transform_32e9a332.js.map
