+++ crates/pack-tests/tests/snapshot/basic/foreign_jsx_transform/output/main.js
@@ -1,5 +1,5 @@
 (globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([
     typeof document === "object" ? document.currentScript : undefined,
-    {"otherChunks":["_project__basic_foreign_jsx_transform_32e9a332.js"],"runtimeModuleIds":["[project]/basic/foreign_jsx_transform/input/index.jsx [client] (ecmascript)"]}
+    {"otherChunks":["_project__basic_foreign_jsx_transform_5da2b40c.js"],"runtimeModuleIds":["[project]/basic/foreign_jsx_transform/input/index.tsx [client] (ecmascript)"]}
 ]);
 // Dummy runtime
