+++ crates/pack-tests/tests/snapshot/typescript/verbatim_module_syntax/config.json
