+++ crates/pack-tests/tests/snapshot/typescript/verbatim_module_syntax/input/index.tsx
@@ -1,0 +1,12 @@
+// This is the opposite of basic/foreign_jsx_transform test case.
+import React from 'react';
+
+const App = () => {
+    return (
+        <div>
+            Verbatim module syntax Test Case
+        </div>
+    )
+}
+
+export default App;
