+++ crates/pack-tests/tests/snapshot/typescript/verbatim_module_syntax/output/input_index_tsx_0a80628f.js
@@ -1,0 +1,18 @@
+(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([typeof document === "object" ? document.currentScript : undefined,
+"[project]/typescript/verbatim_module_syntax/input/index.tsx [client] (ecmascript)", ((__turbopack_context__) => {
+"use strict";
+
+// This is the opposite of basic/foreign_jsx_transform test case.
+const App = ()=>{
+    return /*#__PURE__*/ React.createElement("div", null, "Verbatim module syntax Test Case");
+};
+var __TURBOPACK__default__export__ = App;
+__turbopack_context__.s([
+    "default",
+    0,
+    __TURBOPACK__default__export__
+]);
+}),
+]);
+
+//# sourceMappingURL=input_index_tsx_0a80628f.js.map
