+++ crates/pack-tests/tests/snapshot/typescript/verbatim_module_syntax/output/main.js
@@ -1,0 +1,5 @@
+(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([
+    typeof document === "object" ? document.currentScript : undefined,
+    {"otherChunks":["input_index_tsx_0a80628f.js"],"runtimeModuleIds":["[project]/typescript/verbatim_module_syntax/input/index.tsx [client] (ecmascript)"]}
+]);
+// Dummy runtime
