+++ crates/pack-tests/tests/snapshot/typescript/verbatim_module_syntax/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "compilerOptions": {
+        "verbatimModuleSyntax": false
+    }
+}
